requirements.txt
# Core Frameworks
langchain>=0.3.0
langchain-core>=1.0.0
langchain-community>=0.3.0
streamlit>=1.28.0

# Integration & Models
langchain-groq
langchain-huggingface
transformers>=5.0.0
huggingface-hub>=1.3.0

# Authentication & UI
st-login-form
st-supabase-connection

# Database & Logic
supabase
python-dotenv
pydantic>=2.0.0
langgraph
PyPDF2
# Analytics & Visualization
plotly>=5.0.0
pandas>=1.0.0